+++ tests/test_stream.py
@@ -412,17 +412,11 @@ class TestStanzaStream(StanzaStreamTestBase):
         self.stream.stop()
 
     def test_enqueue_stanza_validates_stanza(self):
-        iq = make_test_iq(type_="get")
+        iq = unittest.mock.Mock()
 
-        with unittest.mock.patch.object(iq, "validate") as validate:
-            self.stream.enqueue_stanza(iq)
+        self.stream.enqueue_stanza(iq)
 
-        self.assertSequenceEqual(
-            [
-                unittest.mock.call(),
-            ],
-            validate.mock_calls
-        )
+        iq.validate.assert_called_with()
 
     def test_start_stop(self):
         self.stream.start(self.xmlstream)